pyproject.toml
[project]
name = "expenses-chatbot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "sqlalchemy",
    "python-dotenv",
    "google-genai",
    "pandas",
    "pydantic-settings>=2.10.1",
]
